+++ src/lib.rs
@@ -4,21 +4,15 @@
 //! # WebAssembly Runtime
 
 mod service;
+mod trace;
 
-use std::any::Any;
-use std::collections::HashMap;
 use std::fs::{self, File};
 use std::io::Write;
 use std::path::PathBuf;
 
 use anyhow::{Result, anyhow};
-use bytes::Bytes;
-use wasmtime::component::{Component, InstancePre, Linker};
-use wasmtime::{Config, Engine, StoreLimits};
-use wasmtime_wasi::{
-    IoView, OutputStream, ResourceTable, StreamError, StreamResult, WasiCtx, WasiCtxBuilder,
-    WasiView,
-};
+use wasmtime::component::{Component, Linker};
+use wasmtime::{Config, Engine};
 
 #[cfg(feature = "http")]
 use crate::service::http;
@@ -32,6 +26,7 @@ use crate::service::keyvalue;
 use crate::service::messaging;
 #[cfg(feature = "rpc")]
 use crate::service::rpc;
+use crate::service::{Ctx, Service};
 
 /// Compile `wasm32-wasip2` component.
 ///
@@ -82,25 +77,6 @@ fn serialize(wasm: &PathBuf) -> Result<Vec<u8>> {
     let engine = Engine::new(&config)?;
     let component = Component::from_file(&engine, wasm)?;
     component.serialize()
-}
-
-/// Service represents a particular runtime service depended on by wasm
-/// components. For example, an HTTP server or a message broker.
-#[async_trait::async_trait]
-pub trait Service: Sync + Send {
-    /// Returns the wasi namespace the service supports. For example, `wasi:http`.
-    fn namespace(&self) -> &'static str;
-
-    /// Add the service to the wasm component linker.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the service encounters an error adding generated
-    /// bindings to the linker.
-    fn add_to_linker(&self, linker: &mut Linker<Ctx>) -> Result<()>;
-
-    /// Start and run the runtime.
-    async fn start(&self, pre: InstancePre<Ctx>) -> Result<()>;
 }
 
 /// Runtime for a wasm component.
@@ -212,132 +188,5 @@ impl Runtime {
         }
 
         Ok(())
-    }
-}
-
-pub type Metadata = Box<dyn Any + Send>;
-
-/// Ctx implements messaging host interfaces. In addition, it holds the host-defined
-/// state used by the wasm runtime [`Store`].
-#[allow(clippy::struct_field_names)]
-pub struct Ctx {
-    table: ResourceTable,
-    ctx: WasiCtx,
-    pub limits: StoreLimits,
-    pub metadata: HashMap<String, Metadata>,
-}
-
-impl Default for Ctx {
-    fn default() -> Self {
-        let mut ctx = WasiCtxBuilder::new();
-        ctx.inherit_args();
-        ctx.inherit_env();
-        ctx.inherit_stdin();
-        ctx.stdout(Stdout {});
-        ctx.stderr(Errout {});
-
-        Self {
-            table: ResourceTable::default(),
-            ctx: ctx.build(),
-            limits: StoreLimits::default(),
-
-            // TODO: wrap Hashmap in custom type to create accessors
-            metadata: HashMap::default(),
-        }
-    }
-}
-
-impl Ctx {
-    /// Create a new Ctx instance.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
-
-impl IoView for Ctx {
-    fn table(&mut self) -> &mut ResourceTable {
-        &mut self.table
-    }
-}
-
-// Implement the [`wasmtime_wasi::ctx::WasiView`] trait for Ctx.
-impl WasiView for Ctx {
-    fn ctx(&mut self) -> &mut WasiCtx {
-        &mut self.ctx
-    }
-}
-
-// Implement debug tracing for Guests by capturing stdout.
-struct Stdout;
-
-impl wasmtime_wasi::StdoutStream for Stdout {
-    fn stream(&self) -> Box<dyn OutputStream> {
-        Box::new(StdoutStream {})
-    }
-
-    fn isatty(&self) -> bool {
-        false
-    }
-}
-
-struct StdoutStream;
-
-#[async_trait::async_trait]
-impl wasmtime_wasi::Pollable for StdoutStream {
-    async fn ready(&mut self) {}
-}
-
-impl OutputStream for StdoutStream {
-    fn write(&mut self, bytes: Bytes) -> StreamResult<()> {
-        let out = String::from_utf8(bytes.to_vec())
-            .map_err(|e| StreamError::LastOperationFailed(anyhow!(e)))?;
-        println!("{out}");
-        Ok(())
-    }
-
-    fn flush(&mut self) -> StreamResult<()> {
-        Ok(())
-    }
-
-    fn check_write(&mut self) -> StreamResult<usize> {
-        Ok(1024 * 1024)
-    }
-}
-
-// Implement error tracing for Guests by capturing stderr.
-struct Errout;
-
-impl wasmtime_wasi::StdoutStream for Errout {
-    fn stream(&self) -> Box<dyn OutputStream> {
-        Box::new(ErroutStream {})
-    }
-
-    fn isatty(&self) -> bool {
-        false
-    }
-}
-
-struct ErroutStream;
-
-#[async_trait::async_trait]
-impl wasmtime_wasi::Pollable for ErroutStream {
-    async fn ready(&mut self) {}
-}
-
-impl wasmtime_wasi::OutputStream for ErroutStream {
-    fn write(&mut self, bytes: Bytes) -> StreamResult<()> {
-        let out = String::from_utf8(bytes.to_vec())
-            .map_err(|e| StreamError::LastOperationFailed(anyhow!(e)))?;
-        println!("{out}");
-        Ok(())
-    }
-
-    fn flush(&mut self) -> StreamResult<()> {
-        Ok(())
-    }
-
-    fn check_write(&mut self) -> StreamResult<usize> {
-        Ok(1024 * 1024)
     }
 }
